refactor(cli): move link command to postgres directory
