+++ packages/cli/src/bin.ts
@@ -33,9 +33,9 @@ import { DebugInfo } from './DebugInfo'
 import { Format } from './Format'
 import { Generate } from './Generate'
 import { Init } from './Init'
-import { Link } from './link/Link'
 import { Mcp } from './mcp/MCP'
 import { Platform } from './platform/_Platform'
+import { Link as PostgresLink } from './postgres/link/Link'
 import { PostgresCommand } from './postgres/PostgresCommand'
 import { Status } from './Status'
 import { Studio } from './Studio'
@@ -114,7 +114,7 @@ async function main(): Promise<number> {
         seed: DbSeed.new(),
       }),
       postgres: PostgresCommand.new({
-        link: Link.new(),
+        link: PostgresLink.new(),
       }),
       generate: Generate.new(),
       version: Version.new(),
