+++ packages/cli/src/postgres/link/Link.ts
@@ -9,9 +9,9 @@ import type { ManagementApiClient } from '@prisma/management-api-sdk'
 import { AuthError, createManagementApiClient } from '@prisma/management-api-sdk'
 import { bold, dim, green, red } from 'kleur/colors'
 
-import { login } from '../management-api/auth'
-import { createAuthenticatedManagementAPI } from '../management-api/auth-client'
-import { FileTokenStorage } from '../management-api/token-storage'
+import { login } from '../../management-api/auth'
+import { createAuthenticatedManagementAPI } from '../../management-api/auth-client'
+import { FileTokenStorage } from '../../management-api/token-storage'
 import { formatCompletionOutput } from './completion-output'
 import { isAlreadyLinked, writeLocalFiles } from './local-setup'
 import { createDevConnection, LinkApiError, listDatabases, listProjects, sanitizeErrorMessage } from './management-api'
